+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/Solvers.java
@@ -10,18 +10,5 @@ import java.util.Map;
 
 public class Solvers {
 
-  private static final Map<String, Integer> priorities = Priorities.build(
-    "SolverCached", "SolverSystem", "SolverDocker", "SolverLocalDB", "SolverCachedRemote"
-  );
-
-  public static List<Solver> sorted(Collection<Solver> source) {
-    final var solvers = new ArrayList<>(source);
-    solvers.sort(Solvers.comparator());
-    return solvers;
-  }
-
-  public static Comparator<Solver> comparator() {
-    return Comparator.comparing(it -> Priorities.compare(priorities, it.name()));
-  }
 
 }
